langchain_bot/requirements.txt
langchain==0.3.4
langchain-openai==0.2.5
openai==1.55.0
python-dotenv==1.0.1
requests==2.32.3

